requirements.txt
anthropic>=0.39.0
openai>=1.50.0
google-genai>=1.0.0
edge-tts>=6.1.0
pyttsx3>=2.90
python-dotenv>=1.0.0
rich>=13.7.0
prompt-toolkit>=3.0.0
